+++ __docs__/wonder-blocks-form/checkbox-group.stories.tsx
@@ -9,7 +9,7 @@ import {LabelLarge, LabelXSmall} from "@khanacademy/wonder-blocks-typography";
 import {Choice, CheckboxGroup} from "@khanacademy/wonder-blocks-form";
 import packageConfig from "../../packages/wonder-blocks-form/package.json";
 
-import ComponentInfo from "../../.storybook/components/component-info";
+import ComponentInfo from "../components/component-info";
 
 export default {
     title: "Packages / Form / CheckboxGroup",