+++ packages/wonder-blocks-modal/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @khanacademy/wonder-blocks-modal
 
+## 7.0.3
+
+### Patch Changes
+
+-   Updated dependencies [7516b239]
+    -   @khanacademy/wonder-blocks-core@11.1.0
+    -   @khanacademy/wonder-blocks-breadcrumbs@3.0.5
+    -   @khanacademy/wonder-blocks-icon-button@6.0.5
+    -   @khanacademy/wonder-blocks-layout@3.0.5
+    -   @khanacademy/wonder-blocks-typography@3.0.5
+
+## 7.0.2
+
+### Patch Changes
+
+-   11a0f5c6: No functional changes. Adding prepublishOnly script.
+-   Updated dependencies [11a0f5c6]
+    -   @khanacademy/wonder-blocks-breadcrumbs@3.0.4
+    -   @khanacademy/wonder-blocks-icon-button@6.0.4
+    -   @khanacademy/wonder-blocks-typography@3.0.4
+    -   @khanacademy/wonder-blocks-theming@3.0.1
+    -   @khanacademy/wonder-blocks-layout@3.0.4
+    -   @khanacademy/wonder-blocks-timing@6.0.1
+    -   @khanacademy/wonder-blocks-tokens@3.0.1
+    -   @khanacademy/wonder-blocks-core@11.0.1
+
+## 7.0.1
+
+### Patch Changes
+
+-   Updated dependencies [d23c9c5f]
+    -   @khanacademy/wonder-blocks-core@11.0.0
+    -   @khanacademy/wonder-blocks-breadcrumbs@3.0.3
+    -   @khanacademy/wonder-blocks-icon-button@6.0.3
+    -   @khanacademy/wonder-blocks-layout@3.0.3
+    -   @khanacademy/wonder-blocks-typography@3.0.3
+
 ## 7.0.0
 
 ### Major Changes